Make uidvalidity field 64-bit
But not in the summary yet; I'll get to that shortly and we aren't even
storing it there anyway right now -- we're always writing zero.

+++ camel/providers/imapx/camel-imapx-server.h
@@ -96,7 +96,7 @@ struct _CamelIMAPXServer {
 	CamelFolderChangeInfo *changes;
 	CamelFolder *select_pending;
 	guint32 permanentflags;
-	guint32 uidvalidity;
+	guint64 uidvalidity;
 	guint32 unseen;
 	guint32 exists;
 	guint32 recent;
